worked on functionalities of useradd,approved,pending,rejected,policy history,user and policy module

+++ user.php
@@ -34,10 +34,12 @@ $res=mysqli_query($con,'SELECT * FROM users where  role=2 order by id desc');
                                     <tr>
                                             <th width="5%" class="serial">#SNO</th>
                                             <th width="5%" class="jonah">ID</th>
-                                            <th width="20%">Name</th>
-                                            <th width="15%">Email</th>
-                                            <th width="15%">Mobile</th>
-                                            <th width="40%">Action</th>
+
+                                            <th width="10%">First Name</th>
+                                            <th width="10%">Last Name</th>
+                                            <th width="10%">Email</th>
+                                            <th width="10%">Mobile</th>
+                                            <th width="30%">Action</th>
 
                                     </tr>
                                     <?php
@@ -67,13 +69,15 @@ $res=mysqli_query($con,'SELECT * FROM users where  role=2 order by id desc');
                                         <tr>
                                             <td ><?php echo $i;  ?></td>
                                             <td ><?php echo $row['id'];  ?></td>
-                                            <td><?php echo $row['name'];  ?></td>
+                                            
+                                            <td><?php echo $row['firstname'];  ?></td>
+                                            <td><?php echo $row['lastname'];  ?></td>
                                             <td><?php echo $row['email'];  ?></td>
                                             <td><?php echo $row['mobile'];  ?></td>
                                             <td>
                                                 <a class="editbtn" href="view_single_user.php?id=<?php echo $row['id'];  ?> "> View</a>
                                                 <a class="editbtn" href="add_user.php?id=<?php echo $row['id'];  ?> "> Edit</a> 
-                                                <a class="deletebtn" href="user.php?id=<?php echo $row['id'];  ?> &type=delete">Delete</a></td>
+                                                <a class="deletebtn" href="user.php?id=<?php echo $row['id'];  ?> &type=delete" onclick="return confirm('Are you sure to delete?')" >Delete</a></td>
                                         </tr>
                                         <?php
                                      $i++; }?>
